requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# HTTP клиент
httpx>=0.24.0

# Переменные окружения
python-dotenv>=1.0.0

# БД (Supabase)
supabase>=2.0.0

# x402gate.io (оплата USDC на Base/EVM)
x402[evm]>=0.1.0

# Pyrogram (Client API для доступа к аккаунту)
pyrogram>=2.0.0
tgcrypto>=1.2.0

# QR-код для /connectqr
qrcode>=7.0

# Тестирование
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
